# backend/openapi/openapi.yaml
openapi: "3.0.3"
info:
  title: Team Manager API
  description: >-
    REST API for the Teamverwaltung sports-club management application.


    Every endpoint may return `429 Too Many Requests` (global per-IP rate
    limit, `RATE_LIMIT_RPS`) or `413 Payload Too Large` (request body cap,
    4 MB by default) — see the `TooManyRequests`/`PayloadTooLarge` reusable
    responses. These aren't repeated on every operation below to keep the
    spec readable; `/auth/login` documents its own stricter rate limit
    explicitly since it differs from the global default.
  version: "1.0.0"
  contact:
    name: Team Manager
    url: https://github.com/yoadey/team-manager
servers:
  - url: /api/v1
    description: Current server
security:
  - bearerAuth: []

# ─── Tags ────────────────────────────────────────────────────────────────────
...
paths:
...
  /teams:
    get:
      tags: [teams]
      operationId: listTeams
      summary: List teams the authenticated user belongs to
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamForUser"
    post:
      tags: [teams]
      operationId: createTeam
      summary: Create a new team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamForUser"
  /teams/{teamId}:
    parameters:
      - $ref: "#/components/parameters/teamId"
    get:
      tags: [teams]
      operationId: getTeam
      summary: Get a single team
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [teams]
      operationId: updateTeam
      summary: Update team settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeamRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /teams/{teamId}/photo:
    parameters:
      - $ref: "#/components/parameters/teamId"
    get:
      tags: [teams]
      operationId: getTeamPhoto
      summary: Get team photo
      responses:
        "200":
          description: JPEG image
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [teams]
      operationId: uploadTeamPhoto
      summary: Upload team photo (scaled to max 800×800 px)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [photo]
              properties:
                photo:
                  type: string
                  format: binary
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
    delete:
      tags: [teams]
      operationId: deleteTeamPhoto
      summary: Remove team photo (revert to icon)
      responses:
        "204":
          description: No Content
  /teams/{teamId}/logo:
    parameters:
      - $ref: "#/components/parameters/teamId"
    get:
      tags: [teams]
      operationId: getTeamLogo
      summary: Get team logo
      responses:
        "200":
          description: JPEG image
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [teams]
      operationId: uploadTeamLogo
      summary: Upload team logo (scaled to max 800×800 px)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [logo]
              properties:
                logo:
                  type: string
                  format: binary
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
    delete:
      tags: [teams]
      operationId: deleteTeamLogo
      summary: Remove team logo (revert to icon)
      responses:
        "204":
          description: No Content
  /teams/{teamId}/invite:
    parameters:
      - $ref: "#/components/parameters/teamId"
    post:
      tags: [teams]
      operationId: createInvite
      summary: Generate a 7-day invite link
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invite"
  /invites/{code}/accept:
    parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
    post:
      tags: [teams]
      operationId: acceptInvite
      summary: Redeem an invite code, adding the authenticated user to its team
      description: >-
        Idempotent: redeeming a code for a team the caller already belongs to
        just returns that team rather than erroring.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcceptInviteResponse"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Members ─────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    teamId:
      name: teamId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    NotFound:
      description: Not Found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    PayloadTooLarge:
      description: >-
        Payload Too Large. All request bodies are capped at 4 MB by default;
        image upload endpoints additionally enforce their own lower limit.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 Problem Details
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
    PermLevel:
      type: string
      enum: [none, read, write]
    Permissions:
      type: object
      description: Per-module permission levels
      required: [events, members, finances, news, polls, settings]
      properties:
        events:
          $ref: "#/components/schemas/PermLevel"
        members:
          $ref: "#/components/schemas/PermLevel"
        finances:
          $ref: "#/components/schemas/PermLevel"
        news:
          $ref: "#/components/schemas/PermLevel"
        polls:
          $ref: "#/components/schemas/PermLevel"
        settings:
          $ref: "#/components/schemas/PermLevel"
...
    Team:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        short:
          type: string
        icon:
          type: string
        iconBg:
          type: string
        iconFg:
          type: string
        description:
          type: string
        hasPhoto:
          type: boolean
        hasLogo:
          type: boolean
        reasonVisibilityRoleIds:
          type: array
          items:
            type: string
            format: uuid
    TeamForUser:
      allOf:
        - $ref: "#/components/schemas/Team"
        - type: object
          required: [myRoles, myPerms, membershipId, memberCount]
          properties:
            myRoles:
              type: array
              items:
                $ref: "#/components/schemas/Role"
            myPerms:
              $ref: "#/components/schemas/Permissions"
            membershipId:
              type: string
              format: uuid
            memberCount:
              type: integer
    AcceptInviteResponse:
      allOf:
        - $ref: "#/components/schemas/TeamForUser"
        - type: object
          required: [alreadyMember]
          properties:
            alreadyMember:
              type: boolean
              description: >-
                True when the caller was already a member of this team before
                redeeming the code (the request was a no-op join-wise).
    CreateTeamRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        icon:
          type: string
        iconBg:
          type: string
        iconFg:
          type: string
    UpdateTeamRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        short:
          type: string
          maxLength: 50
        icon:
          type: string
        iconBg:
          type: string
        iconFg:
          type: string
        description:
          type: string
          maxLength: 10000
        reasonVisibilityRoleIds:
          type: array
          items:
            type: string
            format: uuid
    Invite:
      type: object
      required: [id, teamId, code, link, createdAt, expiresAt]
      properties:
        id:
          type: string
          format: uuid
        teamId:
          type: string
          format: uuid
        code:
          type: string
        link:
          type: string
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time

    # ── Members & Roles ──────────────────────────────────────────────────────
    Role:
      type: object
      required: [id, teamId, name, system, permissions]
      properties:
        id:
          type: string
          format: uuid
        teamId:
          type: string
          format: uuid
        name:
          type: string
        system:
          type: boolean
        color:
          type: string
        permissions:
          $ref: "#/components/schemas/Permissions"
...